api-server/kafka/requirements.txt
psycopg2-binary
dotenv
confluent_kafka
python-dotenv
